Feature/implement token provider with auto refresh on expiry (#72)
* Implement token provider + refactor API client to disconnect and wait for reconnect if the API call failed due to expired token + change reconnection policy to reconnect instantly first few times + reduced default reconnection internals

* Change reconnection logic to have few instant attempts + Add TokenProviderException

* Add simple tool to toggle Debug mode for the SDK

+++ Assets/Plugins/StreamChat/Core/Exceptions/StreamApiException.cs
@@ -10,8 +10,8 @@ namespace StreamChat.Core.Exceptions
     /// </summary>
     public class StreamApiException : Exception
     {
-        public double? StatusCode { get; }
-        public double? Code { get;  }
+        public int? StatusCode { get; }
+        public int? Code { get;  }
         public string Duration { get;  }
         public string ErrorMessage { get;  }
         public string MoreInfo { get;  }
